+++ prqlc/prqlc-parser/src/parser/mod.rs
@@ -19,12 +19,8 @@ mod types;
 // Note that `parse_source` is in `prqlc` crate, not in `prqlc-parser` crate,
 // because it logs using the logging framework in `prqlc`.
 
-pub fn parse_lr_to_pr(
-    source: &str,
-    source_id: u16,
-    lr: Vec<lr::Token>,
-) -> (Option<Vec<pr::Stmt>>, Vec<Error>) {
-    let stream = prepare_stream(lr.into_iter(), source, source_id);
+pub fn parse_lr_to_pr(source_id: u16, lr: Vec<lr::Token>) -> (Option<Vec<pr::Stmt>>, Vec<Error>) {
+    let stream = prepare_stream(lr, source_id);
     let (pr, parse_errors) = stmt::source().parse_recovery(stream);
 
     let errors = parse_errors.into_iter().map(|e| e.into()).collect();
@@ -35,14 +31,15 @@ pub fn parse_lr_to_pr(
 
 /// Convert the output of the lexer into the input of the parser. Requires
 /// supplying the original source code.
-pub(crate) fn prepare_stream(
-    tokens: impl Iterator<Item = lr::Token>,
-    source: &str,
+pub(crate) fn prepare_stream<'a>(
+    tokens: Vec<lr::Token>,
     source_id: u16,
-) -> Stream<lr::TokenKind, Span, impl Iterator<Item = (lr::TokenKind, Span)> + Sized> {
+) -> Stream<'a, lr::TokenKind, Span, impl Iterator<Item = (lr::TokenKind, Span)> + Sized + 'a> {
+    let final_span = tokens.last().map(|t| t.span.end).unwrap_or(0);
+
     // We don't want comments in the AST (but we do intend to use them as part of
     // formatting)
-    let semantic_tokens = tokens.filter(|token| {
+    let semantic_tokens = tokens.into_iter().filter(|token| {
         !matches!(
             token.kind,
             lr::TokenKind::Comment(_) | lr::TokenKind::LineWrap(_)
@@ -52,10 +49,9 @@ pub(crate) fn prepare_stream(
     let tokens = semantic_tokens
         .into_iter()
         .map(move |token| (token.kind, Span::new(source_id, token.span)));
-    let len = source.chars().count();
     let eoi = Span {
-        start: len,
-        end: len + 1,
+        start: final_span,
+        end: final_span,
         source_id,
     };
     Stream::from_iter(eoi, tokens)