+++ prqlc/prqlc/tests/integration/error_messages.rs
@@ -84,11 +84,11 @@ fn test_errors() {
        │                                    ╰── unexpected ’
     ───╯
     Error:
-       ╭─[:1:38]
+       ╭─[:1:39]
        │
      1 │ Mississippi has four S’s and four I’s.
-       │                                      ┬
-       │                                      ╰── Expected * or an identifier, but didn't find anything before the end.
+       │                                       │
+       │                                       ╰─ Expected * or an identifier, but didn't find anything before the end.
     ───╯
     "###);