+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 **Features**:
 
+- Add a option to the experimental documentation generator to output the docs in
+  HTML format. The option is given using the `--format=html` option.
+  (@vanillajonathan, 4791)
+
 **Fixes**:
 
 **Documentation**: